clean_bay_data/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
pandas
numpy
fsspec
gcsfs